requirements.txt
numpy
tensorflow-gpu>=2.1
tensorflow-addons
tensorflow-probability
foolbox
pickle-mixin
